\name{tsne-package}
\Rdversion{1.1}
\alias{tsne-package}

\docType{package}
\title{The tsne-package for multidimensional scaling}
\description{
This package contains one function called \link[tsne]{tsne} which contains all the functionality.
}

\details{
\tabular{ll}{
Package: \tab tsne\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2010-02-19\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}
 
}
\author{
Justin Donaldson
https://github.com/jdonaldson/rtsne
Maintainer: Justin Donaldson (jdonaldson@gmail.com)
}
\references{
L.J.P. van der Maaten and G.E. Hinton. Visualizing High-Dimensional Data Using t-SNE. \emph{Journal of Machine Learning Research} 9 (Nov) : 2579-2605, 2008. 

L.J.P. van der Maaten. Learning a Parametric Embedding by Preserving Local Structure. In \emph{Proceedings of the Twelfth International Conference on Artificial Intelligence and Statistics} (AISTATS), JMLR W&CP 5:384-391, 2009.
}

\keyword{ package }
% \seealso{
% ~~ Optional links to other man pages, e.g. ~~
% ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
% }
% \examples{
% ~~ simple examples of the most important functions ~~
% }
