% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsne.R
\name{tsne_rep}
\alias{tsne_rep}
\title{Best t-SNE Result From Multiple Initializations}
\usage{
tsne_rep(nrep = 10, ...)
}
\arguments{
\item{nrep}{Number of repeats.}

\item{...}{Arguments to apply to each \code{\link{tsne}} run.}
}
\value{
The \code{\link{tsne}} result with the lowest final cost.
If \code{ret_extra} is not \code{FALSE}, then the final costs for all
\code{nrep} runs are also included in the return value list as a vector
called \code{all_costs}.
}
\description{
Run t-SNE multiple times from a random initialization, and return the
embedding with the lowest cost.
}
\details{
This function ignores any value of \code{Y_init} you set, and uses
\code{Y_init = "rand"}.
}
\examples{
\dontrun{
# Return best result out of five random initializations
tsne_iris_best <- tsne_rep(nrep = 5, iris, perplexity = 50, ret_extra = TRUE)
# How much do the costs vary between runs?
range(tsne_iris_best$all_costs)
# Display best embedding found
plot(iris_best$Y)
}
}
